+++ otter/run/run_autograder/constants.py
@@ -2,6 +2,7 @@
 Default configurations for autograding
 """
 
+<<<<<<< HEAD
 from ...utils import convert_config_description_dict
 
 # a dictionary for documenting each configuration and its default value; the dictionary of actual
@@ -70,12 +71,12 @@ DEFAULT_OPTIONS_WITH_DESCRIPTIONS = [
     {
         "key": "filtering",
         "description": "whether the generated PDF should have cells filtered out",
-        "default": True,
+        "default": False,
     },
     {
         "key": "pagebreaks",
         "description": "whether the generated PDF should have pagebreaks between filtered sectios",
-        "default": True,
+        "default": False,
     },
     {
         "key": "debug",
@@ -125,3 +126,33 @@ DEFAULT_OPTIONS_WITH_DESCRIPTIONS = [
 ]
 
 DEFAULT_OPTIONS = convert_config_description_dict(DEFAULT_OPTIONS_WITH_DESCRIPTIONS)
+=======
+DEFAULT_OPTIONS = {
+    "score_threshold": None,
+    "points_possible": None,
+    "show_stdout": False,
+    "show_hidden": False,
+    "show_all_public": False,
+    "seed": None,
+    "grade_from_log": False,
+    "serialized_variables": {},
+    "public_multiplier": 0,
+    "pdf": False,
+    "token": None,
+    "course_id": 'None',
+    "assignment_id": 'None',
+    "filtering": False,
+    "pagebreaks": False,
+    "debug": False,
+    "autograder_dir": '/autograder',
+    "lang": 'python',
+    "miniconda_path": '/root/miniconda3',
+    "test_visibility": 'hidden',
+    "separate_tests": True,
+    "plugins": [],
+    "logo": True,
+    "print_summary": True,
+    "print_score": True,
+    "zips": False,
+}
+>>>>>>> a37ea19f0e054838ff551eecd9073973d5b677f4